feat(admin): add admin layout with sidebar and pages
Create admin layout with Sidebar + Breadcrumb. Add dashboard, dossiers,
suivi, and parametres pages. Restructure /admin to redirect to
/admin/dashboard. Keep existing /admin/users page.
