+++ app/admin/dashboard/page.tsx
@@ -1,0 +1,40 @@
+import { auth } from '@/utils/serverAuth';
+
+export default async function AdminDashboardPage() {
+  const session = await auth();
+
+  return (
+    <div className="mx-auto max-w-6xl">
+      <h1 className="mb-8 text-3xl font-bold">Tableau de bord</h1>
+
+      <div className="grid grid-cols-1 gap-6 md:grid-cols-2 lg:grid-cols-3">
+        <div className="card bg-base-100 shadow-xl">
+          <div className="card-body">
+            <h2 className="card-title">Bienvenue</h2>
+            <p className="text-base-content/70">
+              Connecte en tant que {session?.user?.name}
+            </p>
+          </div>
+        </div>
+
+        <div className="card bg-base-100 shadow-xl">
+          <div className="card-body">
+            <h2 className="card-title">Dossiers</h2>
+            <p className="text-base-content/70">
+              Statistiques globales des dossiers
+            </p>
+          </div>
+        </div>
+
+        <div className="card bg-base-100 shadow-xl">
+          <div className="card-body">
+            <h2 className="card-title">Utilisateurs</h2>
+            <p className="text-base-content/70">
+              Gerer les utilisateurs et les invitations
+            </p>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
