+++ app/admin/dossiers/page.tsx
@@ -1,0 +1,17 @@
+export default function AdminDossiersPage() {
+  return (
+    <div className="mx-auto max-w-6xl">
+      <div className="mb-6 flex items-center justify-between">
+        <h1 className="text-3xl font-bold">Dossiers</h1>
+      </div>
+
+      <div className="card bg-base-100 shadow-xl">
+        <div className="card-body">
+          <p className="text-base-content/70">
+            La liste des dossiers sera affichee ici.
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
