+++ app/admin/parametres/page.tsx
@@ -1,0 +1,15 @@
+export default function AdminParametresPage() {
+  return (
+    <div className="mx-auto max-w-6xl">
+      <h1 className="mb-8 text-3xl font-bold">Parametres</h1>
+
+      <div className="card bg-base-100 shadow-xl">
+        <div className="card-body">
+          <p className="text-base-content/70">
+            Les parametres d&apos;administration seront disponibles ici.
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
