+++ app/admin/suivi/page.tsx
@@ -1,0 +1,15 @@
+export default function AdminSuiviPage() {
+  return (
+    <div className="mx-auto max-w-6xl">
+      <h1 className="mb-8 text-3xl font-bold">Suivi dossier</h1>
+
+      <div className="card bg-base-100 shadow-xl">
+        <div className="card-body">
+          <p className="text-base-content/70">
+            Le suivi detaille des dossiers sera affiche ici.
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
